// LibrarySystem_FrontV1.1/src/app/models/borrowing-two.models.ts
export class BorrowingTwoModel {
    id: string;
    isDelete: any;
    createTime: number;
    operator: any;
    updateTime: number;
    code: string;
    studentName: string;
    professionName: string;
    classesName: string;
    author: string;
    bookName: string;
    name:string;
    content: string;
    borrowTime: number;
    endDate: number;
    realDate: number;
    isOverdue: number;
    renew: number;
    borrowStatus: number;
    searchNum: string;
    picture: string;
    message: string;
    isbn:string;
    location:string;
    remark:string;
    userId:string;
}